+++ druid-shell/src/backend/web/window.rs
@@ -510,6 +510,10 @@ impl WindowHandle {
         warn!("WindowHandle::set_always_on_top unimplemented for web");
     }
 
+    pub fn set_mouse_pass_through(&self, _mouse_pass_thorugh: bool) {
+        warn!("WindowHandle::set_mouse_pass_through unimplemented for web");
+    }
+
     pub fn get_position(&self) -> Point {
         warn!("WindowHandle::get_position unimplemented for web.");
         Point::new(0.0, 0.0)
@@ -522,6 +526,10 @@ impl WindowHandle {
     pub fn get_size(&self) -> Size {
         warn!("WindowHandle::get_size unimplemented for web.");
         Size::new(0.0, 0.0)
+    }
+
+    pub fn is_foreground_window(&self) -> bool {
+        true
     }
 
     pub fn content_insets(&self) -> Insets {