+++ druid-shell/src/window.rs
@@ -225,6 +225,11 @@ impl WindowHandle {
         self.0.set_always_on_top(always_on_top);
     }
 
+    /// Sets whether the mouse passes through the window to whatever is behind.
+    pub fn set_mouse_pass_through(&self, mouse_pass_through: bool) {
+        self.0.set_mouse_pass_through(mouse_pass_through);
+    }
+
     /// Sets where in the window the user can interact with the program.
     ///
     /// This enables irregularly shaped windows. For example, you can make it simply
@@ -242,6 +247,12 @@ impl WindowHandle {
     /// [display points]: crate::Scale
     pub fn set_input_region(&self, region: Option<Region>) {
         self.0.set_input_region(region)
+    }
+
+    /// Returns true if the window is the foreground window or this is unknown.
+    /// Returns false if a different window is known to be the foreground window.
+    pub fn is_foreground_window(&self) -> bool {
+        self.0.is_foreground_window()
     }
 
     /// Returns the position of the top left corner of the window.